7. Bot de Telegram 10-06-2022/requirements.txt
python-telegram-bot
python-dotenv
telegram
numpy